Creo que está arreglado ya el archivo .properties

+++ JavaTracer/src/com/javatracer/model/configuration/Configuration.java
@@ -1,6 +1,7 @@
 package com.javatracer.model.configuration;
 import java.io.File;
 import java.io.FileInputStream;
+import java.io.FileNotFoundException;
 import java.io.FileOutputStream;
 import java.io.FileWriter;
 import java.io.IOException;
@@ -22,42 +23,53 @@ public class Configuration {
     	/*
     	 * Create a file configuration
     	 */
-
-        this.properties = new Properties();
-        try {
-        	/*
-        	 * The file is created on disk
-        	 */
-        	
-        		System.out.println("FICHERO NO EXISTE ");
-        		fileWriter = new FileWriter(CONFIG_FILE_NAME);
-            	
+    	this.properties = new Properties();
+    	
+    	File f = new File(CONFIG_FILE_NAME);
+    	if (f.exists()){
+    		try {
+				this.properties.load(new FileInputStream(CONFIG_FILE_NAME));
+			} catch (FileNotFoundException e) {
+				e.printStackTrace();
+			} catch (IOException e) {
+				e.printStackTrace();
+			}
+    	}
+    	else{	
+    		try {
             	/*
-            	 * The file is uploaded to a File to use the Load method
+            	 * The file is created on disk
             	 */
-            	fileInput = new FileInputStream(CONFIG_FILE_NAME);
             	
-            	/*
-            	 * With the method load, we can load the object properties
-            	 */
-                properties.load(fileInput);
-                /*
-                 * Now we add information to the properties through a Hashmap
-                 */
-           
-                	properties.putAll(getConfig());
-                /*
-                 * Save the data on the properties
-                 */
-                properties.store(new FileOutputStream("configuration.properties"),"Config");
-                fileWriter.close();
-                fileInput.close();
-        	
-        	System.out.println("FICHERO  EXISTE ");
-        	
-        } catch (IOException ex) {
-            ex.printStackTrace();
-        }
+            		fileWriter = new FileWriter(CONFIG_FILE_NAME);
+                	
+                	/*
+                	 * The file is uploaded to a File to use the Load method
+                	 */
+                	fileInput = new FileInputStream(CONFIG_FILE_NAME);
+                	
+                	/*
+                	 * With the method load, we can load the object properties
+                	 */
+                    properties.load(fileInput);
+                    /*
+                     * Now we add information to the properties through a Hashmap
+                     */
+               
+                    	properties.putAll(getConfig());
+                    /*
+                     * Save the data on the properties
+                     */
+                    properties.store(new FileOutputStream("configuration.properties"),"Config");
+                    fileWriter.close();
+                    fileInput.close();
+            	
+            } catch (IOException ex) {
+                ex.printStackTrace();
+            }
+    		
+    	}
+        
         
     }
